+++ .credo.exs
@@ -1,0 +1,15 @@
+%{
+  configs: [
+    %{
+      name: "default",
+      files: %{
+        included: ["lib/", "test/"],
+        excluded: [~r"/_build/", ~r"/deps/"]
+      },
+      strict: true,
+      checks: [
+        {Credo.Check.Readability.ModuleDoc, false}
+      ]
+    }
+  ]
+}